# Conflicts:
#	fleximove-backend/rating-service/src/main/java/com/fleximove/rating/repository/RatingProviderRepository.java
#	fleximove-backend/rating-service/src/main/java/com/fleximove/rating/repository/RatingVehicleRepository.java
#	fleximove-backend/rating-service/src/main/java/com/fleximove/rating/service/RatingService.java

+++ fleximove-backend/vehicle-service/src/main/java/de/fleximove/vehicle/service/services/VehicleService.java
@@ -1,5 +1,6 @@
 package de.fleximove.vehicle.service.services;
 
+import de.fleximove.vehicle.service.clients.BookingServiceClient;
 import de.fleximove.vehicle.service.clients.RatingServiceClient;
 import de.fleximove.vehicle.service.clients.UserServiceClient;
 import de.fleximove.vehicle.service.domain.Vehicle;
@@ -8,8 +9,11 @@ import de.fleximove.vehicle.service.dto.*;
 import de.fleximove.vehicle.service.repository.VehicleRepository;
 import de.fleximove.vehicle.service.utils.DistanceUtils;
 import de.fleximove.vehicle.service.utils.VehicleMapper;
+import feign.FeignException;
 import jakarta.persistence.EntityNotFoundException;
+import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -24,14 +28,16 @@ public class VehicleService {
     private final GeocodingService geocodingService;
     private final UserServiceClient userServiceClient;
     private final RatingServiceClient ratingServiceClient;
+    private final BookingServiceClient bookingServiceClient;
 
     @Autowired
-    VehicleService(VehicleRepository vehicleRepository, VehicleMapper vehicleMapper, GeocodingService geocodingService, UserServiceClient userServiceClient, RatingServiceClient ratingServiceClient){
+    VehicleService(VehicleRepository vehicleRepository, VehicleMapper vehicleMapper, GeocodingService geocodingService, UserServiceClient userServiceClient, RatingServiceClient ratingServiceClient, BookingServiceClient bookingServiceClient){
         this.vehicleRepository = vehicleRepository;
         this.vehicleMapper = vehicleMapper;
         this.geocodingService = geocodingService;
         this.userServiceClient = userServiceClient;
         this.ratingServiceClient = ratingServiceClient;
+        this.bookingServiceClient = bookingServiceClient;
     }
 
     public void registerNewVehicle(VehicleRequest request, Long providerId) {
@@ -43,16 +49,49 @@ public class VehicleService {
         return vehicleRepository.findById(id);
     }
 
+    @Transactional
     public void deleteVehicle(Long id) {
-        //TODO: löschen nur dann möglich, wenn es weder IN_USE noch BOOKED ist
         Vehicle vehicle = fetchVehicleById(id)
                 .orElseThrow(() -> new EntityNotFoundException("Vehicle with ID " + id + " not found"));
+
+        if (vehicle.getStatus() == VehicleStatus.IN_USE || vehicle.getStatus() == VehicleStatus.BOOKED) {
+            throw new IllegalStateException("Vehicle cannot be deleted while it is in use or booked.");
+        }
+
+        ratingServiceClient.deleteRatingsByVehicleId(vehicle.getId());
+        bookingServiceClient.deleteBookingsByVehicleId(vehicle.getId());
+
         vehicleRepository.delete(vehicle);
     }
 
+    @Transactional
     public void deleteAllVehiclesByProviderId(Long providerId) {
-        //delete only if all vehicles are not in use
-        //for available vehicles change the status to RETIRED
+        List<Vehicle> vehicles = vehicleRepository.findAllByProviderId(providerId);
+        if (vehicles.isEmpty()) {
+            throw new EntityNotFoundException("No vehicles found for provider ID: " + providerId);
+        }
+
+        boolean hasActiveVehicles = vehicles.stream()
+                .anyMatch(v -> v.getStatus() == VehicleStatus.IN_USE || v.getStatus() == VehicleStatus.BOOKED);
+
+        if (hasActiveVehicles) {
+            throw new IllegalStateException("Some vehicles are in use or booked. Deletion stopped.");
+        }
+        for (Vehicle v : vehicles) {
+            try {
+                ratingServiceClient.deleteRatingsByVehicleId(v.getId());
+            } catch (FeignException e) {
+                throw new IllegalStateException("Failed to delete ratings for vehicle " + v.getId() + ": " + e.getMessage());
+            }
+        }
+
+        for (Vehicle v: vehicles) {
+            try {
+                bookingServiceClient.deleteBookingsByVehicleId(v.getId());
+            } catch (FeignException e) {
+                throw new IllegalStateException("Failed to delete bookings for vehicle " + v.getId() + ": " + e.getMessage());
+            }
+        }
         vehicleRepository.deleteAllByProviderId(providerId);
     }
 
@@ -133,7 +172,6 @@ public class VehicleService {
 
 
     public void updateVehicleStatus(Long vehicleId, VehicleStatus newStatus) {
-        //TODO: update nur dann möglich, wenn es weder IN_USE noch BOOKED ist
         Vehicle vehicle = fetchVehicleById(vehicleId)
                 .orElseThrow(() -> new EntityNotFoundException("Vehicle not found with ID: " + vehicleId));
         vehicle.setStatus(newStatus);
@@ -148,13 +186,18 @@ public class VehicleService {
     }
 
     public void editVehicleInformation(Long vehicleId, EditVehicleRequest request) {
-        //TODO: update nur dann möglich machen, wenn der Status vom Vehicle weder BOOKED noch IN_USE ist
         Vehicle vehicleToUpdate = fetchVehicleById(vehicleId)
                 .orElseThrow(() -> new EntityNotFoundException("Vehicle not found with ID: " + vehicleId));
+
+        if (vehicleToUpdate.getStatus() == VehicleStatus.BOOKED || vehicleToUpdate.getStatus() == VehicleStatus.IN_USE) {
+            throw new IllegalStateException("Vehicle cannot be updated while it is booked or in use.");
+        }
 
         updateIdentificationNumberIfPresent(vehicleToUpdate, request);
         updateVehicleModelIfPresent(vehicleToUpdate, request);
         updateVehicleTypeIfPresent(vehicleToUpdate, request);
+        updateStatusIfPresent(vehicleToUpdate, request);
+        updateAddressIfPresent(vehicleToUpdate, request);
         updateRestrictionsIfPresent(vehicleToUpdate, request);
         updatePriceIfPresent(vehicleToUpdate, request);
 
@@ -168,7 +211,7 @@ public class VehicleService {
             String newIdent = request.getIdentificationNumber();
             Optional<Vehicle> existingVehicleWithIdentNumber = vehicleRepository.findByIdentificationNumberIdentNumber(newIdent);
             if (existingVehicleWithIdentNumber.isPresent()) {
-                throw new IllegalArgumentException("Identification number already in use.");
+                throw new DataIntegrityViolationException("Identification number already in use.");
             }
 
             vehicle.setIdentificationNumber(new IdentificationNumber(newIdent));
@@ -189,31 +232,40 @@ public class VehicleService {
         }
     }
 
+    private void updateStatusIfPresent(Vehicle vehicle, EditVehicleRequest request) {
+        if (request.getStatus() != null) {
+            VehicleStatus status = VehicleStatus.valueOf(request.getStatus().toUpperCase());
+            vehicle.setStatus(status);
+        }
+    }
+
     private void updatePriceIfPresent(Vehicle vehicle, EditVehicleRequest request) {
         BillingModel billingModel = BillingModel.valueOf(request.getBillingModel().toUpperCase());
         vehicle.setVehiclePrice(new Price(request.getPriceAmount(), billingModel));
     }
 
-    private void updateRestrictionsIfPresent(Vehicle vehicle, EditVehicleRequest request) {
-        if (request.getMinAge() != null || request.getMaxBookingTimeMinutes() != null ||
-                request.getMaxDistanceKm() != null || request.getMaxPassengers() != null ||
-                request.getRequiredLicenseType() != null) {
-
-            VehicleRestrictions restrictions = vehicle.getRestrictions();
-
-            if (request.getMinAge() != null) { restrictions.setMinAge(request.getMinAge()); }
-            if (request.getMaxBookingTimeMinutes() != null) { restrictions.setMaxBookingTimeMinutes(request.getMaxBookingTimeMinutes()); }
-            if (request.getMaxDistanceKm() != null) { restrictions.setMaxDistanceKm(request.getMaxDistanceKm()); }
-            if (request.getMaxPassengers() != null) { restrictions.setMaxPassengers(request.getMaxPassengers()); }
-
-            if (request.getRequiredLicenseType() != null) {
-                restrictions.setRequiredLicense(
-                        DriverLicenseType.licenseTypeFromCode(request.getRequiredLicenseType())
-                );
-            }
-
-            vehicle.setRestrictions(restrictions);
+    private void updateAddressIfPresent(Vehicle vehicle, EditVehicleRequest request) {
+        if (request.getAddress() != null && !request.getAddress().isBlank()) {
+            Location location = geocodingService.geocodeAddress(request.getAddress());
+            vehicle.setCurrentLocation(location);
         }
+    }
+
+
+    private void updateRestrictionsIfPresent(Vehicle vehicle, EditVehicleRequest request) {
+        VehicleRestrictions restrictions = vehicle.getRestrictions();
+
+        if (request.getMinAge() != null) { restrictions.setMinAge(request.getMinAge()); }
+        if (request.getMaxBookingTimeMinutes() != null) { restrictions.setMaxBookingTimeMinutes(request.getMaxBookingTimeMinutes()); }
+        if (request.getMaxDistanceKm() != null) { restrictions.setMaxDistanceKm(request.getMaxDistanceKm()); }
+        if (request.getMaxPassengers() != null) { restrictions.setMaxPassengers(request.getMaxPassengers()); }
+        if (request.getRequiredLicenseType() != null) {
+            restrictions.setRequiredLicense(
+                    DriverLicenseType.licenseTypeFromCode(request.getRequiredLicenseType())
+            );
+        }
+
+        vehicle.setRestrictions(restrictions);
     }
 
     private void validateVehicleConsistency(Vehicle vehicle) {
